WT-13380 Move WT_LOG_DESC inside src/log
Create an internal_types.h header. This is intended to be packaged along
side an internal_api.h header that will be auto-generated. These intern
will be placed in a log_internal.h for convenience.

Well that's the idea atm.
